// src/generated/models/ProductImage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductImage = {
    id: string;
    key: string | null;
    url: string;
    alt?: string | null;
    isPrimary: boolean;
    position: number;
    productId: string;
    variantId?: string | null;
    createdAt: string;
};