// src/generated/models/ProductImageCreateRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductImageCreateRequest = {
    key: string;
    url: string;
    alt?: string | null;
    isPrimary?: boolean;
    position?: number;
    variantId?: string | null;
};