// src/generated/models/ProductImageUpdateRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductImageUpdateRequest = {
    url?: string;
    alt?: string | null;
    isPrimary?: boolean;
    position?: number;
    variantId?: string | null;
};